requirements.txt
# 核心套件
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1

# 日期處理
python-dateutil==2.8.2
pytz==2024.1

# 通知整合
slack-sdk==3.26.1

# 測試工具
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# 開發工具
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# 型別檢查
types-requests==2.31.0.20240218
types-python-dateutil==2.8.19.20240106

# 日誌增強
colorlog==6.8.0

# 系統監控
psutil==5.9.8 